# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YounGeek HackU Kosen 2022 - OpenAPI 3.0
  description: |-
    HackU Kosen 2022 でYounGeekが実装するAPIスキーマ
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: our HackU Kosen 2022 Repository
  url: https://github.com/youngeek-0410/hacku-kosen-2022
servers:
  - url: http://localhost:8080
paths:
  /api/projects:
    post:
      tags:
        - project
      requestBody:
        description: Post name of receiver in project
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
  /api/projects/{project_id}:          
    get:
      tags:
        - project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: text_message_limit
          description: 取得するテキストメッセージの数
          in: query
          required: true
          schema:
            type: integer
        - name: image_message_limit
          description: 取得する画像メッセージの数
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProjectResponse'  
  /api/projects/{project_id}/message:
    post:
      tags:
        - message
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Post new message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
        required: true
      responses:
        '200':
          description: Successful operation
  /api/projects/{project_id}/spotify_music:
    put:
      description: 指定のSpotifyMusicを指定のProjectに登録する
      tags:
        - spotify_music
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Registered Spotify Music data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSpotifyMusicReqest'
        required: true
      responses:
        '200':
          description: Successful operation
  /api/projects/{project_id}/top_text:
    put:
      description: 指定のトップテキストをProjectに登録する
      tags:
        - project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Registered Spotify Music data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTopTextRequest'
        required: true
      responses:
        '200':
          description: Successful operation
  /api/projects/{project_id}/top_image:
    put:
      description: 指定のトップ画像をProjectに登録する
      tags:
        - project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Registered Spotify Music data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTopImageRequest'
        required: true
      responses:
        '200':
          description: Successful operation
  /api/projects/{project_id}/text_messages:
    get:
      tags:
        - message
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTextMessagesResponse'
  /api/projects/{project_id}/image_messages:
    get:
      tags:
        - message
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetImageMessagesResponse'  
                
  /api/projects/{project_id}/publish:
    post:
      tags:
        - project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishProjectResponse' 
        '400':
          description: |
           公開設定が失敗した時。具体的には以下のケース。
           - top_text, top_image, spotify_music が登録されていない
           - text_message, image_message がそれぞれ5件未満の時
  /api/projects/all_id/:
    get:
      tags:
        - project
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllProjectIdResponse'

components:
  schemas:
    CreateProjectRequest:
      type: object
      required: 
        - receiver_name
      properties:
        receiver_name:
          type: string
          description: デジタル色紙を渡す相手の名前
          maxLength: 36
          example: 山田 太郎
    CreateProjectResponse:
      type: object
      required: 
        - project_id
      properties:
        project_id:
          type: string
          description: 作成されたプロジェクトのID
          example: a38ioje
    GetProjectResponse:
      type: object
      required:
        - project_id
        - receiver_name
      properties:
        project_id:
          type: string
          description: プロジェクトのID
          example: a38ioje
        receiver_name:
          type: string
          description: デジタル色紙を渡す相手の名前
          maxLength: 36
          example: 山田 太郎
        is_publish:
          type: boolean
          description: Projectが公開されたかどうか
        spotify_music:
          $ref: "#/components/schemas/SpotifyMusic"
        top_text:
          type: string
          description: 閲覧面ページの一番上に表示する文章。他のメッセージよりも大きく表示される。
          maxLength: 36
          example: "3年間お疲れ様でした！"
        top_image:
          type: object
          description: 閲覧面ページの一番上に表示する画像。他の画像とは独立して表示される。
          properties:
            url:
              type: string
              description: 画像のURL
              example: https://
            width:
              type: integer
              description: 画像の幅(px)
              example: 1200
            height:
              type: integer
              description: 画像の高さ(px)
              example: 600
        text_messages:
          type: object
          properties:
            count:
              type: integer
              description: メッセージ件数
              example: 1
            items:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string
                    description: メッセージ本文
                    example: おっつー
                  sender_name:
                    type: string
                    description: 送信者
                    example: 山田 圭祐
        image_messages:
          type: object
          properties:
            count:
              type: integer
              description: メッセージ件数
              example: 1
            items:
              type: array
              items:
                type: object
                properties:
                  url:
                    type: string
                    description: 画像URL
                    example: https://
                  width:
                    type: integer
                    description: 画像の幅(px)
                    example: 1200
                  height:
                    type: integer
                    description: 画像の高さ(px)
                    example: 600
                  sender_name:
                    type: string
                    description: 送信者
                    example: 山田 圭祐
    CreateMessageRequest:
      type: object
      required: 
        - text
        - sender_name
      properties:
        text:
          type: string
          description: 送信するテキストメッセージ
          maxLength: 256
          example: おっつー
        images:
          type: string
          description: 送信するすべての画像をzip化してさらにBase64エンコードした文字列
          example: 全画像をzip化してさらにBASE64エンコードした文字列
        sender_name:
          type: string
          description: 送信者の名前
          maxLength: 36
          example: 山田 圭祐
    RegisterSpotifyMusicReqest:
      $ref: "#/components/schemas/SpotifyMusic"
    SpotifyMusic:
      type: object
      properties:
        uri:
          type: string
          description: 楽曲を識別するURI
          example: "uri:ljadjoihjgsadih"
    GetTextMessagesResponse:
      type: object
      properties:
        count:
          type: integer
          description: すべてのテキストの件数
          example: 10
        items:
          type: array
          items:
            type: object
            properties:
              text: 
                type: string
                description: テキストメッセージの本文
                example: おつかれ
              sender_name:
                type: string
                description: テキストメッセージ送信者
                example: 山田 花子
    GetImageMessagesResponse:
      type: object
      properties:
        count:
          type: integer
          description: すべての画像の件数
          example: 15
        items:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                description: 画像のURL
                example: https://
              width:
                type: integer
                description: 画像の幅(px)
                example: 1200
              height:
                type: integer
                description: 画像の高さ(px)
                example: 600
              sender_name:
                type: string
                description: 送信者
    RegisterTopTextRequest:
      type: object
      properties:
        top_text:
          type: string
          description: 閲覧面ページの一番上に表示する文章。他のメッセージよりも大きく表示される。
          maxLength: 36
          example: "3年間お疲れ様でした！"
    RegisterTopImageRequest:
      type: object
      properties:
        image:
          type: string
          description: 閲覧面ページの一番上に表示する画像をBase64でエンコードした文字列
          example: BASE64エンコードした文字列くん
    PublishProjectResponse:
      type: object
      properties:
        publication_url:
          type: string
          description: Projectの完成サイトのURL
          example: "https://cloveeee.site/{project_id}"
    GetAllProjectIdResponse:
      type: object
      properties:
        project_ids:
          type: array
          items:
            type: string
            description: プロジェクトのID
            example: a38ioje